+++ src/prometheus_async/tx/__init__.py
@@ -16,7 +16,7 @@
 Twisted-related functionality.
 """
 
-from ._decorators import time, count_exceptions, track_inprogress
+from ._decorators import count_exceptions, time, track_inprogress
 
 
 __all__ = [